requirements.txt
streamlit>=1.24.0
pillow==10.0.1
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.196
numpy==1.24.3
pandas==2.0.3
opencv-python-headless==4.8.1.78